Move ConsulInternalLock from internal class to independent class (#21730)
* Move ConsulInternalLock from internal class to independent class

* Move ConsulInternalLock from internal class to independent class

* Move ConsulInternalLock from internal class to independent class
